# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voisso API",
    "description": "Voisso marketing automation platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/brands/{brand_id}/campaigns/create/{job_id}/content_plan_items/{content_plan_item_id}": {
      "put": {
        "tags": [
          "Brand Campaign Creation Content Plan Items"
        ],
        "summary": "Update Content Plan Item",
        "operationId": "brand_campaign_creation_content_plan_items_update_content_plan_item",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "content_plan_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Plan Item Id"
            }
          },
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestContentPlanItemUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentPlanItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Brand Campaign Creation Content Plan Items"
        ],
        "summary": "Remove Content Plan Item",
        "operationId": "brand_campaign_creation_content_plan_items_remove_content_plan_item",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "content_plan_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Plan Item Id"
            }
          },
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/brands/{brand_id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Get",
        "operationId": "brands_get",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Update",
        "operationId": "brands_update",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_brands_update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Delete",
        "operationId": "brands_delete",
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Brand Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_brands_update": {
        "properties": {
          "request_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Data"
          },
          "logo_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo File"
          }
        },
        "type": "object",
        "title": "Body_brands_update"
      },
      "BrandArchetypeName": {
        "type": "string",
        "enum": [
          "INNOCENT",
          "EVERYMAN",
          "HERO",
          "OUTLAW",
          "EXPLORER",
          "CREATOR",
          "RULER",
          "MAGICIAN",
          "LOVER",
          "CAREGIVER",
          "JESTER",
          "SAGE"
        ],
        "title": "BrandArchetypeName"
      },
      "BrandAudience": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "age_range": {
            "$ref": "#/components/schemas/BrandAudienceAgeRange",
            "default": "ANY"
          },
          "gender": {
            "$ref": "#/components/schemas/BrandAudienceGender",
            "default": "ANY"
          },
          "income_range": {
            "$ref": "#/components/schemas/BrandAudienceIncomeRange",
            "default": "ANY"
          },
          "pain_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pain Points"
          },
          "objections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Objections"
          },
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ContentChannelName"
            },
            "type": "array",
            "title": "Channels"
          }
        },
        "type": "object",
        "title": "BrandAudience"
      },
      "BrandAudienceAgeRange": {
        "type": "string",
        "enum": [
          "TEENS",
          "YOUNG_ADULTS",
          "ADULTS",
          "MIDDLE_AGED",
          "SENIORS",
          "ANY"
        ],
        "title": "BrandAudienceAgeRange"
      },
      "BrandAudienceGender": {
        "type": "string",
        "enum": [
          "MALE",
          "FEMALE",
          "ANY"
        ],
        "title": "BrandAudienceGender"
      },
      "BrandAudienceIncomeRange": {
        "type": "string",
        "enum": [
          "LOW_INCOME",
          "MIDDLE_INCOME",
          "UPPER_MIDDLE_INCOME",
          "HIGH_INCOME",
          "ANY"
        ],
        "title": "BrandAudienceIncomeRange"
      },
      "BrandColor": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hex_value": {
            "type": "string",
            "title": "Hex Value"
          }
        },
        "type": "object",
        "required": [
          "name",
          "hex_value"
        ],
        "title": "BrandColor"
      },
...
      "BrandData": {
        "properties": {
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "media_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Urls"
          },
          "colors": {
            "items": {
              "$ref": "#/components/schemas/BrandColor"
            },
            "type": "array",
            "title": "Colors"
          },
          "brand_mission": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Mission"
          },
          "archetype": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandArchetypeName"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\w{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "audiences": {
            "items": {
              "$ref": "#/components/schemas/BrandAudience"
            },
            "type": "array",
            "title": "Audiences"
          },
          "content_pillars": {
            "items": {
              "$ref": "#/components/schemas/ContentPillar"
            },
            "type": "array",
            "title": "Content Pillars"
          },
          "tone_of_voice": {
            "$ref": "#/components/schemas/BrandToneOfVoice"
          },
          "positioning": {
            "$ref": "#/components/schemas/PositioningBrandData"
          }
        },
        "type": "object",
        "title": "BrandData"
      },
...
      "BrandResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandData"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at"
        ],
        "title": "BrandResponse"
      },
      "BrandToneOfVoice": {
        "properties": {
          "formality_level": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Formality Level",
            "default": 1
          },
          "humour_level": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Humour Level",
            "default": 1
          },
          "irreverence_level": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Irreverence Level",
            "default": 1
          },
          "enthusiasm_level": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Enthusiasm Level",
            "default": 1
          },
          "industry_jargon_usage_level": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Industry Jargon Usage Level",
            "default": 1
          },
          "sensory_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sensory Keywords"
          },
          "excluded_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Words"
          },
          "signature_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signature Words"
          },
          "sentence_length_preference": {
            "$ref": "#/components/schemas/SentenceLengthPreference",
            "default": "MEDIUM"
          }
        },
        "type": "object",
        "title": "BrandToneOfVoice"
      },
...
      "ContentChannelName": {
        "type": "string",
        "enum": [
          "INSTAGRAM",
          "LINKEDIN"
        ],
        "title": "ContentChannelName"
      },
      "ContentFormat": {
        "type": "string",
        "enum": [
          "TEXT",
          "TEXT_WITH_SINGLE_IMAGE"
        ],
        "title": "ContentFormat"
      },
...
      "ContentPillar": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "business_goal": {
            "$ref": "#/components/schemas/ContentPillarBusinessGoal",
            "default": "BUILD_TRUST"
          },
          "type": {
            "$ref": "#/components/schemas/ContentPillarType",
            "default": "THOUGHT_LEADERSHIP"
          },
          "funnel_stage": {
            "$ref": "#/components/schemas/ContentPillarFunnelStage",
            "default": "LOYALTY"
          },
          "content_types": {
            "items": {
              "$ref": "#/components/schemas/ContentTypeName"
            },
            "type": "array",
            "title": "Content Types"
          },
          "audience_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audience Ids"
          },
          "hooks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hooks"
          },
          "calls_to_action": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Calls To Action"
          }
        },
        "type": "object",
        "title": "ContentPillar"
      },
      "ContentPillarBusinessGoal": {
        "type": "string",
        "enum": [
          "DRIVE_ENGAGEMENT",
          "INCREASE_CONVERSION",
          "BUILD_TRUST",
          "GENERATE_LEADS",
          "RETENTION"
        ],
        "title": "ContentPillarBusinessGoal"
      },
      "ContentPillarFunnelStage": {
        "type": "string",
        "enum": [
          "AWARENESS",
          "CONSIDERATION",
          "CONVERSION",
          "LOYALTY"
        ],
        "title": "ContentPillarFunnelStage"
      },
      "ContentPillarType": {
        "type": "string",
        "enum": [
          "EDUCATION",
          "PRODUCT_SERVICE",
          "SOCIAL_PROOF",
          "BEHIND_THE_SCENES",
          "ENTERTAINMENT",
          "COMMUNITY",
          "THOUGHT_LEADERSHIP"
        ],
        "title": "ContentPillarType"
      },
      "ContentPlanItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "channel": {
            "$ref": "#/components/schemas/ContentChannelName"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentTypeName"
          },
          "content_format": {
            "$ref": "#/components/schemas/ContentFormat"
          },
          "image_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Urls",
            "default": []
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "content_data": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TextWithSingleImageContentData"
                  },
                  {
                    "$ref": "#/components/schemas/TextOnlyContentData"
                  }
                ],
                "discriminator": {
                  "propertyName": "content_format",
                  "mapping": {
                    "TEXT": "#/components/schemas/TextOnlyContentData",
                    "TEXT_WITH_SINGLE_IMAGE": "#/components/schemas/TextWithSingleImageContentData"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Data"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "description",
          "channel",
          "content_type",
          "content_format",
          "scheduled_at"
        ],
        "title": "ContentPlanItem"
      },
...
      "ContentTypeName": {
        "type": "string",
        "enum": [
          "TUTORIAL",
          "TIP",
          "GUIDE",
          "FRAMEWORK",
          "MYTH",
          "MISTAKE",
          "INDUSTRY_INSIGHT",
          "DEMO",
          "FEATURE_HIGHLIGHT",
          "PRODUCT_USE",
          "BEFORE_AFTER",
          "BENEFIT",
          "COMPARISON",
          "TESTIMONIAL",
          "REVIEW",
          "CASE_STUDY",
          "CLIENT_STORY",
          "RESULT",
          "SCREENSHOT",
          "FOUNDER_STORY",
          "TEAM_HIGHLIGHT",
          "PROCESS",
          "PRODUCT_BUILDING",
          "DAILY_WORK",
          "MEME",
          "TREND",
          "RELATABLE_POST",
          "HUMOR",
          "CULTURAL_COMMENTARY",
          "OPINION",
          "PREDICTION",
          "HOT_TAKE",
          "POLL",
          "QUESTION",
          "DISCUSSION",
          "CHALLENGE"
        ],
        "title": "ContentTypeName"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "cancelled",
          "failed",
          "accepted"
        ],
        "title": "JobStatus"
      },
      "PositioningBrandData": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "points_of_difference": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Points Of Difference"
          },
          "points_of_parity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Points Of Parity"
          },
          "product_description": {
            "type": "string",
            "title": "Product Description",
            "default": ""
          }
        },
        "type": "object",
        "title": "PositioningBrandData"
      },
...
      "RestContentPlanItemUpdateRequest": {
        "properties": {
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "content_data": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TextWithSingleImageContentData"
                  },
                  {
                    "$ref": "#/components/schemas/TextOnlyContentData"
                  }
                ],
                "discriminator": {
                  "propertyName": "content_format",
                  "mapping": {
                    "TEXT": "#/components/schemas/TextOnlyContentData",
                    "TEXT_WITH_SINGLE_IMAGE": "#/components/schemas/TextWithSingleImageContentData"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Data"
          }
        },
        "type": "object",
        "title": "RestContentPlanItemUpdateRequest"
      },
...
      "SentenceLengthPreference": {
        "type": "string",
        "enum": [
          "SHORT",
          "MEDIUM",
          "LONG"
        ],
        "title": "SentenceLengthPreference"
      },
      "TextOnlyContentData": {
        "properties": {
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "content_format": {
            "type": "string",
            "const": "TEXT",
            "title": "Content Format",
            "default": "TEXT"
          }
        },
        "type": "object",
        "required": [
          "caption"
        ],
        "title": "TextOnlyContentData"
      },
...
      "TextWithSingleImageContentData": {
        "properties": {
          "caption": {
            "type": "string",
            "title": "Caption"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "content_format": {
            "type": "string",
            "const": "TEXT_WITH_SINGLE_IMAGE",
            "title": "Content Format",
            "default": "TEXT_WITH_SINGLE_IMAGE"
          }
        },
        "type": "object",
        "required": [
          "caption",
          "image_url"
        ],
        "title": "TextWithSingleImageContentData"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}